refactor-moved edgeconnect validation to v1alpha1 api dir to enable multi-version validation
