+++ pkg/api/v1alpha1/edgeconnect/validation/api_server_test.go
@@ -1,4 +1,4 @@
-package edgeconnect
+package validation
 
 import (
 	"encoding/json"
