+++ pkg/api/v1alpha1/edgeconnect/validation/automation_validation.go
@@ -1,4 +1,4 @@
-package edgeconnect
+package validation
 
 import (
 	"context"
