+++ pkg/api/v1alpha1/edgeconnect/validation/config.go
@@ -1,4 +1,4 @@
-package edgeconnect
+package validation
 
 import (
 	"context"
